fixup "lib/vfs: consistent device and inode enumeration"
Increase width of inodes and devices from 32 to 64 bits.

Issue #1751

+++ repos/os/include/file_system_session/file_system_session.h
@@ -167,9 +167,9 @@ struct File_system::Status
 	 * XXX  add executable bit
 	 */
 
-	file_size_t size;
-	unsigned    mode;
-	unsigned    inode;
+	file_size_t   size;
+	unsigned      mode;
+	unsigned long inode;
 
 	bool is_directory() const { return mode & MODE_DIRECTORY; }
 	bool is_symlink()   const { return mode & MODE_SYMLINK; }
@@ -185,9 +185,10 @@ struct File_system::Control { /* to manipulate the executable bit */ };
 struct File_system::Directory_entry
 {
 	enum Type { TYPE_FILE, TYPE_DIRECTORY, TYPE_SYMLINK };
-	unsigned inode;
-	Type     type;
-	char     name[MAX_NAME_LEN];
+
+	unsigned long inode;
+	Type          type;
+	char          name[MAX_NAME_LEN];
 };
